Reorganize files, split requestCompletion into its own file
